Split mass-assignment eval fixtures
