+++ evals/fixtures/mass-assignment-basic/README.md
@@ -1,9 +1,9 @@
 # Mass-assignment fixture
 
-The member-only `/account` route validates only the JSON shape, then passes
-the full request body to `account.update(body)`. A member can therefore set
-server-owned fields such as `role` or `owner_id`.
+The member-only `/account` route in `account.py` validates only the JSON shape,
+then passes the full request body to `account.update(body)`. A member can
+therefore set server-owned fields such as `role` or `owner_id`.
 
-The `/profile` route is the paired negative case. It copies only an explicit
-allow-list of editable fields and overwrites `owner_id` from the authenticated
-principal, so it must not produce a mass-assignment finding.
+The `/profile` route in `profile.py` is the paired negative case. It copies
+only an explicit allow-list of editable fields and overwrites `owner_id` from
+the authenticated principal, so it must not produce a mass-assignment finding.
